Move dependency resolution machinery to utils
